+++ docs/CHANGELOG.md
@@ -8,7 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.19.35.0329] - 2026-03-29
 
 - Fix Map style dosen't been saved at close
-- Fix OSM tile caching problem GRIB2 and Lightning got their indepndent panels Fixed opacity not loaded propely
+- Fix OSM tile caching problem 
+- GRIB2 and Lightning got their indepndent panels
+- Fixed opacity not loaded propely
 
 ## [1.19.34.0326] - 2026-03-26