+++ docs/CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.19.36.0329] - 2026-03-29
+
+- Expand Credits and License sections with legal notices for Silk.NET, ILGPU, CSJ2K, Earthquakes Canada, and Blitzortung
+- Update CHANGELOG.md
+
 ## [1.19.35.0329] - 2026-03-29
 
 - Fix Map style dosen't been saved at close